backend-services/notification_service/requirements.txt
django
djangorestframework
pymysql
cryptography
